fix: rename component to CardFAQUI, move state inside it
